Added .gitignore

+++ src/Driver.cpp
@@ -7,9 +7,9 @@ using namespace std;
 int main()
 {
     Schema nation("nation", "data/nation.tbl");
-    nation.addAttribute("nationkey", FLOAT);
+    nation.addAttribute("nationkey", LONG);
     nation.addAttribute("name", STRING);
-    nation.addAttribute("regionkey", FLOAT);
+    nation.addAttribute("regionkey", LONG);
     nation.addAttribute("comment", STRING);
 
     cout << nation << endl;
